# ui/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Refactored free-fs with a subject-match-java style layered architecture.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://192.168.164.18:30067"
    }
  ],
...
  "paths": {
    "/share/{shareId}": {
      "get": {
        "summary": "分享页面",
        "deprecated": false,
        "tags": [
          "SharePageController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "renderFileSharePage",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "分享ID",
            "required": true,
            "name": "shareId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "提取码（可选）",
            "required": false,
            "name": "shareCode"
          }
        ]
      }
    },
...
    "/files/{fileId}": {
      "get": {
        "summary": "查询文件详情",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "getFileById",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "文件ID",
            "required": true,
            "name": "fileId"
          }
        ]
      }
    },
...
    "/shares": {
      "post": {
        "summary": "创建分享",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileShareDTO"
                }
              }
            }
          }
        },
        "operationId": "createFileShare",
        "parameters": []
      },
      "get": {
        "summary": "按用户查询分享列表",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultFileShareDTO"
                }
              }
            }
          }
        },
        "operationId": "listFileSharesByUser",
        "parameters": []
      },
      "delete": {
        "summary": "批量取消分享",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "cancelFileShares",
        "parameters": []
      }
    },
    "/shares/pages": {
      "get": {
        "summary": "分页查询分享",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultFileShareDTO"
                }
              }
            }
          }
        },
        "operationId": "pageFileSharesByUser",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "No comments found.",
            "required": true,
            "example": 0,
            "name": "pageIndex"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "No comments found.",
            "required": true,
            "example": 20,
            "name": "pageSize"
          }
        ]
      }
    },
    "/shares/{shareId}": {
      "get": {
        "summary": "查询分享详情",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileShareDTO"
                }
              }
            }
          }
        },
        "operationId": "getFileShareById",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "分享ID",
            "required": true,
            "name": "shareId"
          }
        ]
      }
    },
    "/shares/{shareId}/info": {
      "get": {
        "summary": "获取分享页信息（精简版）",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileShareThinDTO"
                }
              }
            }
          }
        },
        "operationId": "getShareInfo",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "分享ID",
            "required": true,
            "name": "shareId"
          }
        ]
      }
    },
    "/shares/{shareId}/verify-code": {
      "post": {
        "summary": "验证分享提取码",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyShareCodeCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultBoolean"
                }
              }
            }
          }
        },
        "operationId": "verifyShareCode",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "分享ID",
            "required": true,
            "name": "shareId"
          }
        ]
      }
    },
    "/shares/{shareId}/items": {
      "get": {
        "summary": "获取分享文件列表",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultFileRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "listShareFiles",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "分享ID",
            "required": true,
            "name": "shareId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "提取码（可选）",
            "required": false,
            "name": "shareCode"
          }
        ]
      }
    },
    "/shares/{shareId}/download/{fileId}": {
      "get": {
        "summary": "下载分享内文件",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/defaultPrimitive"
                }
              }
            }
          }
        },
        "operationId": "downloadShareFile",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "分享ID",
            "required": true,
            "name": "shareId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "文件ID",
            "required": true,
            "name": "fileId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "提取码（可选）",
            "required": false,
            "name": "shareCode"
          }
        ]
      }
    },
...
    "/shares/{shareId}/access-records": {
      "post": {
        "summary": "记录分享访问",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareAccessRecordCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "recordAccess",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "分享ID",
            "required": true,
            "name": "shareId"
          }
        ]
      },
      "get": {
        "summary": "查询分享访问记录",
        "deprecated": false,
        "tags": [
          "FileShareController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultFileShareAccessRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "listAccessRecords",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "分享ID",
            "required": true,
            "name": "shareId"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MultiResultFileRecordDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileRecordDTO"
            }
          }
        }
      },
      "MultiResultFileShareAccessRecordDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileShareAccessRecordDTO"
            }
          }
        }
      },
...
      "SingleResultFileShareThinDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/FileShareThinDTO"
          }
        }
      },
...
      "PageResultFileShareDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "totalCount": {
            "format": "int64",
            "description": "总数",
            "type": "number",
            "example": 0
          },
          "pageSize": {
            "format": "int32",
            "description": "页数",
            "type": "integer",
            "example": 0
          },
          "pageIndex": {
            "format": "int32",
            "description": "当前页数",
            "type": "integer",
            "example": 0
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileShareDTO"
            }
          }
        }
      },
      "CreateShareAccessRecordCommand": {
        "type": "object",
        "properties": {
          "shareId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "accessIp": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "accessAddress": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "browser": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "os": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "CreateShareCommand": {
        "type": "object",
        "properties": {
          "userId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "fileIds": {
            "description": "No comments found.",
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": "\"\"\",\"\"\""
          },
          "shareName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "expireType": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          },
          "expireTime": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "scope": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "needShareCode": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean",
            "example": true
          },
          "maxViewCount": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          },
          "maxDownloadCount": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          }
        },
        "required": [
          "userId",
          "fileIds"
        ]
      },
...
      "FileShareThinDTO": {
        "type": "object",
        "properties": {
          "shareId": {
            "format": "string",
            "description": "分享ID",
            "type": "string"
          },
          "shareName": {
            "format": "string",
            "description": "分享名称",
            "type": "string"
          },
          "shareCode": {
            "format": "string",
            "description": "提取码",
            "type": "string"
          },
          "fileCount": {
            "format": "int32",
            "description": "文件数量",
            "type": "integer",
            "example": 0
          },
          "hasCheckCode": {
            "format": "boolean",
            "description": "是否需要提取码",
            "type": "boolean",
            "example": true
          },
          "isExpire": {
            "format": "boolean",
            "description": "是否已过期",
            "type": "boolean",
            "example": true
          },
          "expireTime": {
            "format": "string",
            "description": "过期时间",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "viewCount": {
            "format": "int32",
            "description": "已浏览次数",
            "type": "integer",
            "example": 0
          },
          "downloadCount": {
            "format": "int32",
            "description": "已下载次数",
            "type": "integer",
            "example": 0
          }
        }
      },
...
      "defaultPrimitive": {
        "type": "object",
        "properties": {
          "shareIds": {
            "description": "分享ID列表,[array of string]",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SingleResultFileRecordDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/FileRecordDTO"
          }
        }
      },
...
      "VerifyShareCodeCommand": {
        "type": "object",
        "properties": {
          "shareId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "shareCode": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        },
        "required": [
          "shareId",
          "shareCode"
        ]
      },
...
      "SingleResultFileShareDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/FileShareDTO"
          }
        }
      },
...
      "FileRecordDTO": {
        "type": "object",
        "properties": {
          "fileId": {
            "format": "string",
            "description": "文件ID",
            "type": "string"
          },
          "originalName": {
            "format": "string",
            "description": "原始文件名",
            "type": "string"
          },
          "displayName": {
            "format": "string",
            "description": "显示文件名",
            "type": "string"
          },
          "suffix": {
            "format": "string",
            "description": "文件后缀",
            "type": "string"
          },
          "size": {
            "format": "int64",
            "description": "文件大小（字节）",
            "type": "number",
            "example": 0
          },
          "isDir": {
            "format": "boolean",
            "description": "是否为目录",
            "type": "boolean",
            "example": true
          },
          "parentId": {
            "format": "string",
            "description": "父目录ID",
            "type": "string"
          },
          "userId": {
            "format": "string",
            "description": "所属用户ID",
            "type": "string"
          },
          "storageSettingId": {
            "format": "string",
            "description": "存储配置ID",
            "type": "string"
          },
          "isDeleted": {
            "format": "boolean",
            "description": "是否已删除",
            "type": "boolean",
            "example": true
          },
          "isFavorite": {
            "format": "boolean",
            "description": "是否收藏",
            "type": "boolean",
            "example": true
          },
          "uploadTime": {
            "format": "string",
            "description": "上传时间",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "updateTime": {
            "format": "string",
            "description": "更新时间",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          }
        }
      },
...
      "SingleResultBoolean": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "format": "boolean",
            "description": "返回的单条数据",
            "type": "boolean"
          }
        }
      },
...
      "SingleResultVoid": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "type": "object"
          }
        }
      },
...
      "MultiResultFileShareDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileShareDTO"
            }
          }
        }
      },
...
    }
  }
}